Editor: Make `serializeBlocks` a util.
